# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  models.Payment:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      currency:
        type: string
      customer_id:
        type: string
      description:
        type: string
      id:
        type: string
      payment_method:
        type: string
      payment_status:
        $ref: '#/definitions/models.PaymentStatus'
      stripe_id:
        type: string
      updated_at:
        type: string
    type: object
  models.PaymentRequest:
    properties:
      amount:
        type: integer
      currency:
        maxLength: 3
        type: string
      customer_id:
        type: string
      description:
        type: string
      payment_method:
        type: string
      token:
        description: |-
          CardNumber    string `json:"card_number" validate:"required_if=PaymentMethod card,omitempty,credit_card"`
          CardExpMonth  int    `json:"card_exp_month" validate:"required_if=PaymentMethod card,omitempty,min=1,max=12"`
          CardExpYear   int    `json:"card_exp_year" validate:"required_if=PaymentMethod card,omitempty,min=2025"`
          CardCVC       string `json:"card_cvc" validate:"required_if=PaymentMethod card,omitempty,len=3"`
        type: string
    required:
    - amount
    - currency
    - customer_id
    - description
    - payment_method
    - token
    type: object
  models.PaymentResponse:
    properties:
      client_secret:
        type: string
      message:
        type: string
      payment:
        $ref: '#/definitions/models.Payment'
      payment_status:
        type: string
    type: object
  models.PaymentStatus:
    enum:
    - pending
    - succeeded
    - failed
    - refunded
    type: string
    x-enum-varnames:
    - PaymentStatusPending
    - PaymentStatusSucceeded
    - PaymentStatusFailed
    - PaymentStatusRefunded
...
  response.ErrorResponse:
    properties:
      code:
        type: string
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
host: localhost:8085
info:
  contact:
    email: aaravmahajan2003@gmail.com
    name: Aarav Mahajan
    url: https://github.com/aaravmahajanofficial
  description: This is the API server for the Scalable E-commerce Platform. It provides
    endpoints for managing users, products, carts, orders, payments, and notifications.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Scalable E-commerce Platform API
  version: "1.0"
paths:
...
  /payments:
    get:
      description: Retrieves a paginated list of payment records for the authenticated
        user. Requires authentication.
      parameters:
      - description: 'Page number for pagination (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Number of items per page (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of payments
          schema:
            allOf:
            - $ref: '#/definitions/models.PaginatedResponse'
            - properties:
                Data:
                  items:
                    $ref: '#/definitions/models.Payment'
                  type: array
              type: object
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List user's payments with pagination
      tags:
      - Payments
    post:
      consumes:
      - application/json
      description: Creates a payment intent (e.g., with Stripe) for a specified order
        and returns the client secret needed for frontend processing. Requires authentication.
      parameters:
      - description: Payment Request Details (Order ID, Amount, Currency, Customer
          ID)
        in: body
        name: payment
        required: true
        schema:
          $ref: '#/definitions/models.PaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully initiated payment, includes client secret
          schema:
            $ref: '#/definitions/models.PaymentResponse'
        "400":
          description: Validation error or invalid input
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden - Attempting to pay for another user's order
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Order not found or already paid
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error or payment provider error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Initiate a payment for an order
      tags:
      - Payments
  /payments/{id}:
    get:
      description: Retrieves details for a specific payment record. Requires authentication.
        (Authorization might be needed to restrict access).
      parameters:
      - description: Payment ID (Internal DB ID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved payment details
          schema:
            $ref: '#/definitions/models.Payment'
        "400":
          description: Invalid or missing payment ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden - User cannot access this payment" // Add if authz
            implemented
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Payment not found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payment details by ID
      tags:
      - Payments
  /payments/webhook:
    post:
      consumes:
      - application/json
      description: Receives and processes webhook events from Stripe (e.g., payment
        success, failure) to update internal payment and order statuses. This endpoint
        should not require application-level authentication but relies on Stripe's
        signature verification.
      parameters:
      - description: Stripe webhook signature for verification
        in: header
        name: Stripe-Signature
        required: true
        type: string
      - description: Raw Stripe event payload (JSON)
        in: body
        name: payload
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad request (e.g., missing signature, failed reading body,
            invalid payload)
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Webhook signature verification failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error during webhook processing
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Handle incoming Stripe webhooks
      tags:
      - Payments (Internal)
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"